Update logic for fetching JWKS from the core
